+++ src/app/pages/NFTByFtxId/HallOfFameItem/HallOfFameItem.tsx
@@ -1,12 +1,13 @@
 import { Typography } from '@material-ui/core';
 import useTranslation from 'app/hooks/useTranslation';
+import styles from './HallOfFameItem.module.scss';
 
-export const HallOfFameItem = ({ offerPrice, number, styles, totalQuantity }: HallOfFameItemProps) => {
+export const HallOfFameItem = ({ offerPrice, number, totalQuantity }: HallOfFameItemProps) => {
   const t = useTranslation();
 
   return (
-    <div className={styles.mainContent__priceContainer}>
-      <div className={styles.mainContent__priceItem}>
+    <div className={styles.hallOfFameItem__price}>
+      <div className={styles.hallOfFameItem__priceItem}>
         <div>
           <Typography component="div" variant="h6">
             {t('nft.price')}
@@ -14,13 +15,13 @@ export const HallOfFameItem = ({ offerPrice, number, styles, totalQuantity }: Ha
           <Typography component="span">${offerPrice}</Typography>
         </div>
       </div>
-      <div className={styles.mainContent__priceItem}>
+      <div className={styles.hallOfFameItem__priceItem}>
         <Typography component="div" variant="h6">
           {t('nft.editionNumber')}
         </Typography>
         <Typography component="span">#{number}</Typography>
       </div>
-      <div className={styles.mainContent__priceItem}>
+      <div className={styles.hallOfFameItem__priceItem}>
         <Typography component="div" variant="h6">
           {t('nft.totalReleased')}
         </Typography>
@@ -34,5 +35,4 @@ interface HallOfFameItemProps {
   offerPrice: number;
   number: number;
   totalQuantity: number;
-  styles: any;
 }
