+++ src/app/pages/NFTByFtxId/HeroleteItem/HeroleteItem.tsx
@@ -1,28 +1,28 @@
 import { ReactComponent as RibbonIcon } from 'app/assets/icons/ribbon.svg';
 import { ReactComponent as BackgroundIcon } from 'app/assets/icons/background.svg';
-import { Typography } from '@material-ui/core';
-import useTranslation from 'app/hooks/useTranslation';
+import { Button, Typography } from '@material-ui/core';
 import { colors } from 'app/constants/constants';
+import styles from './HeroleteItem.module.scss';
+import useTranslation from 'app/hooks/useTranslation';
 
 export const HeroleteItem = ({
   background,
   offerPrice,
   priceInUsd,
   quoteCurrency,
-  styles,
   tier,
 }: HeroleteItemProps) => {
   const t = useTranslation();
 
   return (
     <>
-      <div className={styles.mainContent__heroleteItem}>
+      <div className={styles.heroleteItem}>
         <div>
           <RibbonIcon fill={colors.white} />
           <Typography component="div" variant="subtitle1">
             Tier
           </Typography>
-          <Typography component="div" variant="subtitle1" className={styles.mainContent__heroleteItemInfo}>
+          <Typography component="div" variant="subtitle1" className={styles.heroleteItem__info}>
             {tier}
           </Typography>
         </div>
@@ -31,12 +31,12 @@ export const HeroleteItem = ({
           <Typography component="div" variant="subtitle1">
             Background
           </Typography>
-          <Typography component="div" variant="subtitle1" className={styles.mainContent__heroleteItemInfo}>
+          <Typography component="div" variant="subtitle1" className={styles.heroleteItem__info}>
             {background}
           </Typography>
         </div>
       </div>
-      <div className={styles.mainContent__heroleteItemPrice}>
+      <div className={styles.heroleteItem__price}>
         <Typography component="div" variant="h6">
           {t('nft.price')}:
         </Typography>
@@ -47,6 +47,11 @@ export const HeroleteItem = ({
           <small>{`($${(priceInUsd * offerPrice).toFixed(2)})`}</small>
         </div>
       </div>
+      <div className={styles.heroleteItem__button}>
+        <Button fullWidth variant="contained">
+          {t('nft.buyButton')}
+        </Button>
+      </div>
     </>
   );
 };
@@ -56,6 +61,5 @@ interface HeroleteItemProps {
   offerPrice: number;
   priceInUsd: number;
   quoteCurrency: string;
-  styles: any;
   tier: string;
 }
